==> spacex capstone (reference)/coursera-basheergit-spacex-data-analysis-project/3_Data_Analysis_with_SQL.ipynb <==
# Data Analysis with SQL

%load_ext sql

import csv, sqlite3
import pandas as pd

%sql sqlite:///my_data1.db

con = sqlite3.connect("my_data1.db")
cur = con.cursor()
df = pd.read_csv("Data_SpaceX_WebScrapped.csv")
df.to_sql("SPACEXTBL", con, if_exists='replace', index=False,method="multi")

%sql create table SPACEXTABLE as select * from SPACEXTBL where Date is not null

%sql select * from SPACEXTABLE limit 4

%sql select distinct [Launch site] from SPACEXTABLE

%sql select * from SPACEXTABLE where [Launch site] like 'CCA%' limit 5

%sql select sum([Payload mass]) from SPACEXTABLE where Customer = 'NASA'

%sql select avg([Payload mass]) from SPACEXTABLE where [Version Booster]= 'F9 v1.1'

%sql select min(Date) from SPACEXTABLE where [Booster landing] = 'Success'

%sql select distinct [Version Booster] from SPACEXTABLE where [Booster landing] = 'Success' and [Payload mass] between 4000 and 6000

%sql select [Launch outcome],count([Launch outcome]) from SPACEXTABLE group by [Launch outcome]
