==> spacex capstone (reference)/coursera-basheergit-spacex-data-analysis-project/5_Data_Visualization_for_Launch_Sites.ipynb <==
# SpaceX Launch Sites Data Visualization

## Mark All Sites on a Map

### Importing Packages

import folium
import pandas as pd

from folium.plugins import MarkerCluster
from folium.plugins import MousePosition
from folium.features import DivIcon

### Importing the data

spacex_df=pd.read_csv("data_falcon9.csv")
spacex_df.head()

# Select relevant sub-columns: `Launch Site`, `Lat(Latitude)`, `Long(Longitude)`, `class`
spacex_df = spacex_df[['LaunchSite', 'Latitude', 'Longitude', 'Class']]
launch_sites_df = spacex_df.groupby(['LaunchSite'], as_index=False).first()
launch_sites_df = launch_sites_df[['LaunchSite', 'Latitude', 'Longitude']]
launch_sites_df

### Creating the Map

# Start location is NASA Johnson Space Center
nasa_coordinate = [29.559684888503615, -95.0830971930759]
# Initial the map
site_map = folium.Map(location=nasa_coordinate,zoom_start=5)
# For each launch site, add a Circle object based on its coordinate (Lat, Long) values. In addition, add Launch site name as a popup label
for index,row in launch_sites_df.iterrows():
    coordinate = [row['Latitude'],row['Longitude']]
    circle = folium.Circle(coordinate, radius=1000, color='#d35400', fill=True).add_child(folium.Popup(row['LaunchSite']))
    marker = folium.map.Marker(
    coordinate,
    icon=DivIcon(
        icon_size=(20,20),
        icon_anchor=(0,0),
        html='<div style="font-size: 12; color:#d35400;"><b>%s</b></div>' % row['LaunchSite'],
        )
    )
    site_map.add_child(circle)
    site_map.add_child(marker)

site_map

## Mark the success/failed launches for each site on the map

### Create markercluster object

marker_cluster = MarkerCluster()

### Create a new column in Launch_Site data frame to store marker color

launch_sites = spacex_df[['LaunchSite', 'Latitude', 'Longitude', 'Class']]
markers = []
for index,record in launch_sites.iterrows(): 
    if record['Class']==1:
        markers.append('green')
    elif record['Class']==0:
        markers.append('red')
launch_sites['marker_color'] = markers
launch_sites.tail()

### For each launch result in `spacex_df` data frame, add a `folium.Marker` to `marker_cluster`

# Add marker_cluster to current site_map
site_map.add_child(marker_cluster)

for index, record in launch_sites.iterrows():
    # TODO: Create and add a Marker cluster to the site map
    coordinate = [record['Latitude'],record['Longitude']]
    marker = folium.map.Marker(
    coordinate,
    icon=folium.Icon(color='white',icon_color=record['marker_color'])
    )
    marker_cluster.add_child(marker)

site_map
